mailer/requirements.txt
email-validator
pymongo
kafka-python